backend/pyproject.toml
[project]
name = "python-django-template"
version = "0.1.0"
description = "Python Django Template"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.35.33",
    "cachetools==5.5.0",
    "celery==5.4.0",
    "celery-types==0.22.0",
    "colorlog==6.8.2",
    "dj_database_url==2.2.0",
    "Django==5.1.1",
    "django-axes[ipware]==7.0.0",
    "django-cors-headers==4.4.0",
    "django-silk==5.2.0",
    "django-split-settings==1.3.2",
    "django-storages==1.14.4",
    "django_redis==5.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework_simplejwt==5.3.1",
    "drf-spectacular==0.27.2",
    "gunicorn==23.0.0",
    "psycopg==3.2.3",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "redis==5.0.8",
    "sentry-sdk==2.14.0",
]

[tool.uv]
dev-dependencies = [
    "mypy==1.11.1",
    "pre-commit==3.8.0",
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "pytest-django==4.8.0",
    "ruff==0.6.2",
]

[tool.uv.workspace]
members = ["user"]

[tool.setuptools]
py-modules = []

[tool.ruff]
show-fixes = true
line-length = 90
exclude = [
    "migrations",
    ".venv",
]

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["T201"]
ignore = [
    "E501",   # Line too long  # Will be fixed by black
    "A003",
    "RUF001",
    "ERA001",
    "F841",   # Local variable is assigned to but never used

    "ANN101",
    "ANN102",  # Missing type annotation for `cls` in classmethod
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in `{name}`

    "D10",

    "D203",
    "D213",
    "EXE002",

    "PD011",
    "UP040",  # Type alias uses `TypeAlias` annotation instead of the `type` keyword
]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",  # Module imported but unused
]
"tests/**" = [
    "S101",  # Use of assert detected
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports_per_module = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
strict = false
implicit_reexport = true
exclude = [
    "venv",
    "migrations",
]

[tool.pytest.ini_options]
minversion = "7.0"
DJANGO_SETTINGS_MODULE = "api.config.settings"
addopts = "--exitfirst -vv --cov --cov-report=html --cov-fail-under=80"
testpaths = [
    "tests",
]
